backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.12.0
httpx==0.27.2
python-dotenv==1.0.1
python-multipart==0.0.9
edge-tts==6.1.12
